delete data from Firebase

+++ src/components/organisms/feed/Feed_post_section.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @next/next/no-img-element */
-import { db } from "@/configs/firebase";
+import { db, storage } from "@/configs/firebase";
 import {
   ChartBarIcon,
   HeartIcon,
@@ -16,6 +16,7 @@ import {
   collection,
   deleteDoc,
 } from "firebase/firestore";
+import { deleteObject, ref } from "firebase/storage";
 
 import React, { useEffect, useState } from "react";
 import Moment from "react-moment";
@@ -43,21 +44,33 @@ export default function Feed_post_section({ post, userInfo }: Props) {
 
   /* sendin like to firestore */
   async function likePost() {
-    if (liked) {
-      await deleteDoc(doc(db, "tweet", post.id, "likes", userInfo?.uid));
+    if (userInfo) {
+      if (liked) {
+        await deleteDoc(doc(db, "tweet", post.id, "likes", userInfo?.uid));
+      } else {
+        await setDoc(doc(db, "tweet", post.id, "likes", userInfo?.uid), {
+          userName: userInfo.displayname,
+        });
+      }
     } else {
-      await setDoc(doc(db, "tweet", post.id, "likes", userInfo?.uid), {
-        userName: userInfo.displayname,
-      });
+    }
+  }
+
+  // deletePost function
+
+  function deletePost() {
+    if (window.confirm("Are you Sure you want to delete?")) {
+      deleteDoc(doc(db, "tweet", post.id));
+      deleteObject(ref(storage, `tweet/${post.id}/image`));
     }
   }
 
   return (
     <div className="flex p-3 cursor-pointer border-b border-gray-200  hover:bg-gray-50">
       {/* user image */}
-      {userInfo?.photoURL && (
+      {post?.data().userImg && (
         <img
-          src={userInfo?.photoURL}
+          src={post?.data().userImg}
           alt="user_image"
           className="h-11  w-11 rounded-full hover:brightness-95 cursor-pointer mr-4"
         />
@@ -71,7 +84,7 @@ export default function Feed_post_section({ post, userInfo }: Props) {
             <div className="flex items-center justify-between">
               <div className="flex items-center space-x-1 whitespace-nowrap">
                 <h4 className="font-bold text-[15px] sm:text-[16px] hover:underline">
-                  {post.data().userName}
+                  {post?.data().userName}
                 </h4>
                 <span className="text-sm sm:text-[15px] text-gray-500">
                   {" "}
@@ -101,7 +114,13 @@ export default function Feed_post_section({ post, userInfo }: Props) {
             {/* icons */}
             <div className="flex items-center justify-between text-gray-500 p-2">
               <ChatIcon className="h-9 w-9 hoverEffect p-2 hover:bg-sky-100 hover:text-sky-500 rounded-full" />
-              <TrashIcon className="h-9 w-9 hoverEffect  p-2  hover:bg-red-100 hover:text-red-500 rounded-full" />
+              {post.data().id === userInfo?.uid && (
+                <TrashIcon
+                  onClick={deletePost}
+                  className="h-9 w-9 hoverEffect  p-2  hover:bg-red-100 hover:text-red-500 rounded-full"
+                />
+              )}
+
               <div className="flex items-center">
                 {liked ? (
                   <HeartIconFilled
@@ -114,7 +133,11 @@ export default function Feed_post_section({ post, userInfo }: Props) {
                     className="h-9 w-9 hoverEffect  p-2 hover:bg-red-100 hover:text-red-500 rounded-full"
                   />
                 )}
-                {likes.length > 0 && <span>{likes.length}</span>}
+                {likes.length > 0 && (
+                  <span className={`${liked && "text-red-600"}`}>
+                    {likes.length}
+                  </span>
+                )}
               </div>
               <ShareIcon className="h-9 w-9 hoverEffect p-2  hover:bg-sky-100 hover:text-sky-500 rounded-full" />
               <ChartBarIcon className="h-9 w-9 hoverEffect  p-2 hover:bg-sky-100 hover:text-sky-500 rounded-full" />
